chore: add build command before any test runs
Added `ctor` as a dependency so that a function to build is triggered before any tests. This doesn't seem to slow things down if the the test contracts aren't updated.

Also made binaryen a feature since the C++ building slowed things down and seemed to break rust analyzer.

Reorganized the tests so that a singe test binary is built instead of individual ones.