Inline variables in format strings
### What
Inline variables in format strings.

### Why
Rust 1.66 introduces a lint to clippy that requires variables to be embedded in format strings.

At this moment the beta (1.66) and nightly (1.67) versions of clippy will emit warnings. This is inconvenient for people choosing to use beta and nightly. In the near future (6 weeks) we'll see this lint in stable.