Add ability to read/write events to sandbox:
The sandbox will now serialize all events that occur during an
invocation of `invoke` to a JSON blob on disk. This blob
corresponds exactly to the JSON RPC response when calling
`getEvents()`.

This enables the `--events-file` flag for the `soroban events`
subcommand, which reads this file from disk, filters based on
the given contract IDs (optional, via `--id`), and outputs then
to stdout.

This also adds the event type (system or contract) to the
output.

Remaining items include:
 - adding support for `DebugEvent`s
 - streaming from the on-disk event store