feat: simplify config consumption
First of all, no longer allow setting `--config-dir`. Only allow using
the workspace config (the default) or the global config by passing
`--global` when adding new configs.

Then, when _consuming_ configs in commands like `invoke`, `read`, etc
(that is, in all parts of the CLI that are not nested under `config`),
don't allow specifying `--global`. In addition to being hard to
interpret (what does it mean to invoke _globally_?), this specificity
introduced potential for user error: if I add a global identity called
"ahalabs", and I call `invoke` with `--identity ahalabs` but forget to
also include `--global`, I would have gotten errors.

Instead, the new code always uses the same behavior when _reading_
configs:

1. Check local config files for an identity/network with the given name
2. If not found, check global config files for that name

This introduces the possibility that a user may add both a local and a
global config with the same name, and could become confused that the
global is not being used. To help prevent such confusion, this also adds
some `println` statements to tell the CLI user which config is being
used, and to warn when adding one with a duplicate name.