refactor: mv arg-parse tests to integration dir
Like the `tests/it/custom_types.rs` tests, these arg-parsing tests rely
on the `target/wasm32-unknown-unknown/test-wasms/test_custom_types.wasm`
file.

But the lookup logic for the `test-wasms` files in the integration test
directory was far superior, throwing a useful message about how you need
to run `make build-test-wasms` if the files are not found. And
unfortunately, unit tests run before integration tests, so the nice
message was never seen, and instead we ended up with errors like:

    thread 'contract::invoke::test::parse_enum' panicked at 'called
    `Result::unwrap()` on an `Err` value: Os { code: 2, kind: NotFound,
    message: "No such file or directory" }',
    cmd/soroban-cli/src/contract/invoke.rs:715:18

I wanted to reuse the nice test-wasms lookup logic, rather than
copy-pasting.

The arg-parsing tests were contained within `invoke.rs`, but they made
no use of anything in the `invoke.rs` file. They were fairly clean to
move, with one big exception: they require importing `strval`!

To import `strval` in the integration tests, I had to add a `lib.rs`, so
that the integration tests can import things directly from the main
library, rather than only using the built binary.

The CLI is written in such a way that it could become a useful library,
in addition to being a binary. Are the current exports the most useful
as part of this library? I doubt it, but they were what I needed to
export in order to move these tests. I'm not sure if it's worth fleshing
out the "library" at this time.