feat!: `--source-account` remove `account`, `secret-key`, `identity` (#459)
* feat!: `--source-account` remove `account`, `secret-key`, `identity`

The sandbox now uses the public key of the associated source account.
Uses prefix to determine whether `seed:`, `id:` or `sk:`.  If none is provided it uses the default test identity which is a seed phrase that is generated by `config identity generate --default-seed ..`.  This means that tests can now use `--hd-path=x` to generate different accounts without needing to use an identity file.

* fix: clippy

* feat: ENV for Soroban Account

* fix: merge issues

* feat: remove need for prefix

* fix: update e2e