fix: make run_in_sandbox pub everywhere it appears
This allows it to be called from tests with something like:

    workspace
        .cmd_arr::<deploy::Cmd>(&["--id", "a", "--wasm", &WASM_A.path().to_string_lossy()])
        .run_in_sandbox()
        .unwrap();

HOWEVER, not all commands contain a `run_in_sandbox` method, which means
this is only a half-measure. Attempting to run tests with commands other
than the ones modified here or `invoke`, which already had a public
`run_in_sandbox`, will require special treatment.