feat(bindings-ts): make Wallet injectable
By default, if you do not provide a `wallet` option to one of the
exported functions or `invoke`, then it will continue to import and use
Freighter.

But now you can also pass your own Wallet interface to any of these
functions.

This facilitates testing the generated libraries with Node, as well as
using them with other Wallet options.