feat(CLI): Each generated contract method adds <contract-arg>-file-path
This allows passing file paths in place of arguments.  Files are read as strings for all types except Bytes and BytesN where the file is read as raw bytes.

The info about the new args is in the help doc header for each subcommand, otherwise the new args are hidden to not clog up the help documentation.