feat(bindings-ts)!: rm --contract-name, add --overwrite
Fixes #752

Removes the `--contract-name` argument from `bindings typescript`, since
this was often confused with `--output-dir` and, really, just using the
file path from the end of `--output-dir` is sufficient.

Adds `--overwrite`. If you specify an `--output-dir` that already
exists, by default it will not be overwritten. Instead, a descriptive error
will be thrown, letting you know that you can add `--overwrite` to
overwrite the directory.

Previously, if you specified an `--output-dir` that already existed, it
would create a new directory with the same name inside this directory,
which can lead to confusing bugs.