feat: no more need for `Method` suffix
When we exported functions for each method, we needed to make sure we
only used safe JS words. This no longer matters, since they're exposed
as methods on the `Contract` class.

Since this change required rebuilding the snapshot, I also made sure the
`readonly` settings were added to the source, rather than manually to
the snapshot.