chore: update ts bindings (#973)
* chore: update ts bindings

What
----

Update the typescript bindings with all the changes that landed in the futurenet-compatible `smartdeploy` branch since `main` stopped working with futurenet.

Why
---

Lots of good improvements and fixes are not yet in `main`, ~including tests for the generated libraries.~ _CI tests will be included in a future PR; reasonable time-sink exceeded for this one.

Closes: https://github.com/stellar/soroban-tools/issues/996

Known limitations
-----------------

The newly-added tests are not yet being exercised in CI; you can see the commit history for all the details of how we tried to work around this. At this point, we will:

- wait for https://github.com/stellar/quickstart/pull/507 to get merged & included in a published quickstart image
- retry the `bindings-ts.yml` workflow, which was almost working except that friendbot was not starting properly when quickstart was run _in CI_ (though it does start properly, even if a bit slowly, when quickstart is run locally)
- after a basic run-tests-with-published-quickstart is working, the next step is to build a custom quickstart image using _this repository's_ RPC version, and use that custom quickstart to run the integration tests. This might be able to be accomplished [as described by ChatGPT here](https://chat.openai.com/share/0f355fbe-93b9-4c8d-9c12-7f54846134bd).

* build: bump delta for flaky test

* build: increase deltas in cli simulation test

---------

Co-authored-by: Willem Wyndham <willem@ahalabs.dev>
Co-authored-by: Tsachi Herman <24438559+tsachiherman@users.noreply.github.com>