feat(CLI): Transaction Handler that handles the state transitions
This renames the `transaction` module to `txn`. Added `txn::Handler` and `txn::State`. The Handler handles the different steps in preparing a transaction to be submitted. From simulation, authentication, and signing.

This is the first step in allowing intermediate transactions be returned and used.
Also in improving error messages from failed transactions.

- [ ] Do not submit transactions that have no auth or writes. We have the data from simulation already.