Remove slop arg
The original reason I added the slop arg was to automatically pass
anything in the slop args directly to `docker run`. But, now that we're
using bollard to handle the container interaction, it seems that we need
to be a bit more specific with how we handle the arguments to docker
run.