Making progress
i needed to look at the js implementation again (hw-app-str) - they have
a note that says to pass the `signatureBase` into the `signTransaction`
fn. I had forgotten about this. The `signatureBase` fn is defined in
js-stellar-base and is the value that should be sent to the network. it
is TransactionSignaturePayload.toXDR(). Which is also in the signer
train and i didn't notice it. 🙈