fix(init): don't gitignore triple-dash
The `---` is only for the `README.md` file, not `gitignore`.

I've also switched the "(ab)used link label (for use with reference
style links)"-style comment (`[//] #`) for an HTML comment (`<!-- ...
-->`). This means it will be rendered in the HTML, which might be
preferable, and is at a minimum benign. I hadn't been familiar with the
`[//]: #`-style comment until I saw it here (and searched to find a
description of it, along with the "(ab)use" line above, [here]), and I
don't think it's worth the potential confusion for users of `soroban
contract init`.

  [here]: https://stackoverflow.com/a/20885980/249801