Allow for retries when checking if the emulator is running (#1459)
The reason we were experiencing intermittent failures with these tests is
because when checking if the emulator was fully up and running,
sometimes the container wouldn't even be started yet, and would panic in
the `wait_for_emulator_start_text` fn. This change adds a retry with
exponential backoff so if the container isn't up yet, and the request
fails, it will try again after waiting.