Remove dependencies to secret.rs and locator.rs from secure_store.rs
In upcoming refactors, secret.rs and locator.rs are going to depend on
secure_store directly, instead of keyring.rs so that keyring.rs can be
entirely behind a feature flag. Without the changes in this commit there
is a cyclical dependency on secret -> secure_store -> secret. Same for
locator.