Remove check for self.asset, offer or data_name before adding LedgerKeyAccount
We were previously checking to see if asset, offer or data-name were
also passed in. If not, the LedgerKeyAccount of the given
account was added to the list of keys. However, there is an edge case,
that if the user passes in 'xlm' as the asset, the entries was returning
empty because there is not an xlm trustline so no key was passed in.

As a solution, this PR always include the LedgerKeyAccount in the list
of keys, so the LedgerEntryAccount is always returned when --account is
passed in as an arg