fix(ts-bindings): valid TS when >1 contracterror (#2031)
We were generating a separate `const Errors` definition for every
`#[contracterror]` in the original contract source. For example, with
the new `custom_type` source code, the generated TS would be:

```ts
export const Errors = {
  /**
   * Please provide an odd number
   */
  1: {message:"NumberMustBeOdd"}
}
export const Errors = {
  /**
   * Some contract libraries contain extra #[contracterror] definitions that end up compiled
   * into the main contract types. We need to make sure tooling deals with this properly.
   */
  100: {message:"HowCouldYou"}
}
```

This now names the `const` after the name of the error enum in the
original file, following the logic of other enum types. It will also
rename `Error` to `Errors`, to avoid conflicts with JS `Error` type. So
the updated `custom_type` source code now generates:

```ts
export const Errors = {
  /**
   * Please provide an odd number
   */
  1: {message:"NumberMustBeOdd"}
}
export const Erroneous = {
  /**
   * Some contract libraries contain extra #[contracterror] definitions that end up compiled
   * into the main contract types. We need to make sure tooling deals with this properly.
   */
  100: {message:"HowCouldYou"}
}
```

These constants are NOT ACTUALLY USED by the TS, but are continuing to
be exported to avoid considering this a breaking change to the library
generation logic.

This is needed in order to generate TS bindings for [OpenZeppelin's
Fungible Token][1]. The contract itself defines [`contracterror` number
1][2], the `pausable` libraries defines [100-series
`contracterror`s][3], and the `fungible` library creates [200-series
`contracterror`s][4].

  [1]: https://github.com/OpenZeppelin/stellar-contracts/blob/main/examples/fungible-token-interface/src/contract.rs
  [2]: https://github.com/OpenZeppelin/stellar-contracts/blob/54eec16f8f8d8373936147464d6eb81eff3c58c2/examples/fungible-token-interface/src/contract.rs#L29C1-L34C2
  [3]: https://github.com/OpenZeppelin/stellar-contracts/blob/54eec16f8f8d8373936147464d6eb81eff3c58c2/packages/contract-utils/pausable/src/pausable.rs#L74-L82
  [4]: https://github.com/OpenZeppelin/stellar-contracts/blob/54eec16f8f8d8373936147464d6eb81eff3c58c2/packages/tokens/fungible/src/fungible.rs#L156-L182

Co-authored-by: Willem Wyndham <willem@ahalabs.dev>