Make envelope the default fetch when no subcommand is given (#2036)
* Make envelope the default fetch when no subcommand is given

In order to flatten an optional field, I needed to remove the
group(skip) on network::Args. But this lead to a conflict because there
are multiple groups with an implicit name of 'Args' that were all trying
to be flattened. The solution to this was add a specific group id to
network::Args.

* Make envelope the default fetch when no subcommand is given

* Refactor to use args impl

* Clean up