Group CLI options into sections in help docs (#2272)
### What
Move the doc-gen binary code into its own crate, separate from the
soroban-cli crate, and have it be dependent on the unmerged pull request
to clap-markdown that adds support for grouping cli options:
- https://github.com/ConnorGray/clap-markdown/pull/48

### Why
The grouped options experience is a massive improvement. It makes the
help documentation easier to scan and understand. It's the way the
options are organised on the command line and the help markdown
documentation should follow suite. While the pull request adds the
feature upstream, the pull request hasn't been accepted yet.

The clap-markdown dependency isn't needed for the stellar-cli other than
doc-gen, so it is in any case good to get the dependency outside of the
soroban-cli crate, so that it isn't downloaded by cargo everytime
someone installs the stellar-cli. Moving doc-gen into its own crate is
trivial because the soroban-cli is designed to be embedded, and achieves
that. Additionally because the doc-gen crate does not need to be
published it can reference the unreleased pull request without issue.