Update rpc client to get_events by passing topics as [TopicFilter] (#2292)
### What

closes https://github.com/stellar/stellar-cli/issues/2021

related rpc-client pr:
https://github.com/stellar/rs-stellar-rpc-client/pull/57

Updates the topic filter structure that is being passed to the rpc
server `getEvents` method, to align with it's expected API.

The GetEventsRequest, from the rpc server, is expecting the topics to be
an array of TopicFilters, and a TopicFilter is an array of
SegmentFilters. So it is expecting an array of arrays.

### Why
We are currently just passing in a segment filter, without nesting it in
an array of arrays, and are seeing an error when doing so.