Remove unused clap imports (#2329)
### What

Remove unused clap::arg and clap::command imports.

### Why
With newer clap versions, the derive macros bring these imports in
automatically. And it looks like the new version of rust, 1.92.0, must
have updated it unused imports check to be more strict so we were seeing
a bunch of failures on build.

### Known limitations

n/a